+++ packages/plugin-packages/mashroom-http-proxy-add-user-headers/src/AddUserHeadersHttpProxyInterceptor.ts
@@ -1,3 +1,4 @@
+import {escapeHeaderValue} from '@mashroom/mashroom-utils/lib/header_utils'
 import type {Request, Response} from 'express';
 import type {IncomingMessageWithContext, MashroomLogger} from '@mashroom/mashroom/type-definitions';
 import type {MashroomSecurityService, MashroomSecurityUser} from '@mashroom/mashroom-security/type-definitions';
@@ -64,7 +65,7 @@ export default class AddUserHeadersHttpProxyInterceptor implements MashroomHttpP
             [this.userNameHeader]: user.username,
         };
         if (user.displayName) {
-            headers[this.displayNameHeader] = user.displayName;
+            headers[this.displayNameHeader] = escapeHeaderValue(user.displayName);
         }
         if (user.email) {
             headers[this.emailHeader] = user.email;